Build PPI Network

Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network.

BuildPPINetwork(
  stringdb.edges = NA,
  gm.network = NA,
  db.filepaths = c(),
  ppi.network.name = "ppi.network"
)

Arguments

stringdb.edges

Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database

gm.network

GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA

db.filepaths

A vector of paths to the additional ppi network files; defaults to an empty vector

ppi.network.name

Desired name for the output protein-protein interaction network using all entered database input; defaults to ppi.network

Value

A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks

Examples

BuildPPINetwork(ex.cccn.matrix, ppi.network.name = "ex.ppi.network")
#> Error in BuildPPINetwork(ex.cccn.matrix, ppi.network.name = "ex.ppi.network"): No data input.
utils::head(ex.ppi.network)
#>     Gene.1   Gene.2 STRINGdb.combined_score   GM.weights
#> 9     MYH9   ABLIM1               0.1240713 0.0005895754
#> 23    MYH9   ATP5PO               0.1426449           NA
#> 25 SLC38A2   ATP5PO               0.1641902           NA
#> 29     APP   ATP5PO               0.2503715 0.0147000789
#> 35  ATP5PO ATP6V0D1               0.5527489           NA
#> 59    ATG3    HSPA9               0.1619614           NA