Combines Spearman dissimilarity and Euclidean distance in parallel

This function uses parallel computing to calculate both Spearman dissimilarity and Euclidean distance, combines them, and performs t-SNE.

CombinedPar(ptmtable.df, ptmtable)

Arguments

ptmtable.df

A data frame containing numeric data.

ptmtable

A dataset for post-translational modifications.

Value

A matrix containing t-SNE coordinates (3D).

Examples

CombinedPar(ptmtable.df, ptmtable)
#>               [,1]         [,2]         [,3]
#>   [1,]  -0.4490388   6.57718653  13.80815148
#>   [2,]  -4.8535749   0.51901207  -1.24007854
#>   [3,]   9.9317251 -10.94849757  -5.20387575
#>   [4,]  -4.9802853   2.27026971   0.60936584
#>   [5,]  -2.3926779   7.54744142  12.97838826
#>   [6,]   6.1728372  -8.83033314  -6.16403871
#>   [7,]   8.1192683 -11.05667443  -5.37168508
#>   [8,]  12.2883101  -0.17952437   9.90600344
#>   [9,] -34.8681121  16.68360268 -20.22394757
#>  [10,]   9.7486022 -20.76874700  -6.20431388
#>  [11,]  13.0900021 -11.71590491  -3.54629652
#>  [12,]  -2.9618393   7.64984647  11.35988257
#>  [13,]   0.8640962   6.01136431  14.51962072
#>  [14,]  11.9158259 -10.42406513  -3.65271513
#>  [15,]   9.4184475  -0.54371021  14.93147364
#>  [16,]  11.0329309 -12.25033946  -4.18864220
#>  [17,]   8.3381544   4.67738118  13.45519828
#>  [18,]   6.9763026   3.68493384  13.08277155
#>  [19,]   2.7871988   5.18376061  14.96204013
#>  [20,]   9.0037914  -1.79579817  13.42503718
#>  [21,]  10.4905697  -2.00392994  14.58506504
#>  [22,]   8.5033955   1.59466212  13.53021355
#>  [23,]   7.0441152   1.55358112  14.74368656
#>  [24,]  -3.3925954   6.27109732  12.31574705
#>  [25,] -35.4938726  14.92863012 -22.33888647
#>  [26,]  11.1810781  -9.16445050  -4.83257783
#>  [27,] -35.2740028  17.11765915 -22.47297200
#>  [28,] -35.7775461  18.23089452 -18.77756116
#>  [29,]   9.5355118 -20.40699912  -4.65566451
#>  [30,]  11.3669312 -11.37658554  -2.41356170
#>  [31,]   8.1895515   3.10253887  17.61177625
#>  [32,]  12.4899127  -0.26850394   7.95134177
#>  [33,]  10.6056321 -12.71776647  -1.54238265
#>  [34,]   4.6003902   3.76815845  15.70495793
#>  [35,] -33.6208761  15.22794073 -21.68081444
#>  [36,] -33.0681528  15.80795256 -19.34309342
#>  [37,]  14.8784845 -10.89137885  -3.08893489
#>  [38,] -36.9678869  16.53435008 -19.26741235
#>  [39,]  14.4841486  -1.09185377   5.65032083
#>  [40,] -33.3910489  17.18548515 -21.23878467
#>  [41,] -33.6975201  17.68143515 -18.96007888
#>  [42,]   5.9703300  -9.82678682  -4.85947905
#>  [43,]  11.3962810  -0.15048195  14.52854588
#>  [44,]  -3.6736285   1.63946174   0.05151657
#>  [45,]  11.0922335  -1.24341215  16.21610131
#>  [46,]   9.0695142  -1.47288206  10.39045037
#>  [47,] -36.7751539  18.05421729 -20.82032710
#>  [48,]   2.6757375   6.02293849  13.03504152
#>  [49,]   4.7419017   4.16187438  13.06568287
#>  [50,]  -0.2553731   5.97095087  10.45126035
#>  [51,]   2.3578342  -4.57903232  -4.63043582
#>  [52,]   0.8868811   4.06752526   9.23165038
#>  [53,]   8.2976183   2.49969280  15.37245527
#>  [54,] -36.3097883  14.83086340 -20.26736277
#>  [55,] -34.7875815  18.68222126 -20.71102278
#>  [56,]  10.8062088 -19.54841075  -6.44167104
#>  [57,]   5.7977110   5.23942754  15.08071631
#>  [58,]   8.3188025  -9.18653272  -4.23522701
#>  [59,]   9.1991244 -18.63834785  -6.02217970
#>  [60,]   6.9979958  -9.24755422  -8.03771702
#>  [61,]  -5.1672095  -0.91485224  -3.30444790
#>  [62,]  12.5365279  -0.17905135  16.01505279
#>  [63,] -37.1361245  16.11714083 -21.74178089
#>  [64,]  -3.6199842  -2.62258205  -4.50790617
#>  [65,] -35.1830327  15.90711484 -18.27003496
#>  [66,]  12.4832533  -3.27315204   1.40662451
#>  [67,]   5.6425987  -6.16049738  -5.21887766
#>  [68,]   6.6943256  -7.96187067  -3.79193711
#>  [69,]   7.4311570  -6.70477432  -4.48400392
#>  [70,]  14.2000351   0.08742981   8.15560591
#>  [71,]  -4.9796244   0.86766435  -3.32279516
#>  [72,]  13.5620500  -1.33016839   7.39918085
#>  [73,]  13.1957422  -1.65883670   4.75764700
#>  [74,]  -0.3502028   4.53989893   9.79872494
#>  [75,]   2.2998888   4.64948699  12.35923518
#>  [76,]  -4.9790585   2.42595113  -1.36893995
#>  [77,]  -1.4143135   5.56457422  11.51896978
#>  [78,]   8.8382273  -8.83919078  -7.76617443
#>  [79,]   7.4485368  -7.38640367  -6.90791236
#>  [80,]  13.0285702  -2.66036203   2.83852852
#>  [81,] -34.4086145  14.49342971 -20.12624211
#>  [82,]  10.8226183   1.96665377  15.92365326
#>  [83,]  14.6079912 -10.37480606  -5.77659302
#>  [84,]  14.6577196 -11.65524378  -5.60382196
#>  [85,]  -3.6979220  -0.15613280  -4.15467371
#>  [86,]  -1.9043749  -0.10044781  -3.63908957
#>  [87,]  -3.3198740   1.48285141  -3.20471711
#>  [88,]   9.3617277  -6.30465934  -1.20933953
#>  [89,]   9.1068318  -7.35497364  -5.63776718
#>  [90,]  13.4557444   0.45706399   6.60066614
#>  [91,]  -1.7426140  -1.90382864  -3.80353671
#>  [92,]  -3.2468912  -0.28782569  -1.84354485
#>  [93,]   8.8401977  -4.43590573  -2.98913098
#>  [94,]   9.6560991  -3.96351983  -2.18861135
#>  [95,]  11.1256059 -20.56916925  -5.18135311
#>  [96,]  10.9459951  -0.10784251  12.65515003
#>  [97,]  -2.9500363   6.11713521  10.14985098
#>  [98,]   9.6894295   2.52695600  14.15431773
#>  [99,]  -3.5186357  -2.27274498  -2.76927969
#> [100,]  10.3128068 -18.67436335  -5.04141852