Get STRINGdb PPI data

This function finds protein-protein interaction weights by consulting utilizing the STRINGdb database.

GetSTRINGdb(
  cccn.matrix,
  STRINGdb.name = "string.edges",
  nodenames.name = "nodenames"
)

Arguments

cccn.matrix

A matrix showing strength of relationships between proteins using common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))

STRINGdb.name

Desired name for the output STRINGdb data frame; defaults to "string.edges"

nodes.name

Desired name for list of gene names; defaults to nodenames

Value

Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database and a list of gene names

Examples

cccn.cfn.tools:::ex.FindPPIEdges(ex.cccn.matrix, "ex.string.edges", "ex.nodenames")
#> Error in cccn.cfn.tools:::ex.FindPPIEdges(ex.cccn.matrix, "ex.string.edges",     "ex.nodenames"): argument "ppi.network.name" is missing, with no default