Make Database Input File

This function outputs a file consisting entirely of gene names, each produced on a new line. This data can be copy and pasted into a database input in order to get protein-protein interaction data.

MakeDBInput(cccn.matrix, file.path.name = "db_nodes.txt")

Arguments

cccn.matrix

A matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))

file.path.name

Path for the output file; defaults to db_nodes.txt

Value

A file with all of the gene names which can be copy and pasted into the GeneMania cytoscape app, data frame of the names of the genes

Examples

#MakeDBInput(ex.nodenames)
cat(ex.nodenames[[1]], sep = '\n')
#> ABLIM1
#> PKP4
#> CTTN
#> MYH9
#> PARD3
#> TMEM192
#> TUBA1B
#> HNRNPU
#> PLEKHA6
#> ATP5PO
#> EZR
#> HSPA1B
#> PI4K2A
#> SLC12A4
#> PSMC1
#> EEF1A1
#> RBM4
#> SF3B2
#> PTPRH
#> SEC16A
#> CTNNA2
#> FRS2
#> OAT
#> RBM14
#> SLC9A1
#> ACOX1
#> ACTR5
#> LCK
#> MYO18A
#> NFS1
#> RPRD2
#> DNM2
#> EMC7
#> HNRNPH2
#> HSPA9
#> OXCT1
#> PHB2
#> RMC1
#> RNASEH2B
#> SUPT16H
#> RASSF8
#> APH1A
#> TJP3
#> ZNF185
#> NSFL1C
#> PTPRK
#> HNRNPF
#> TNK2
#> APP
#> PTTG1IP
#> SGMS2
#> TMEM106B
#> ATG3
#> ATP6V0D1
#> MERTK
#> SYAP1
#> CTNNAL1
#> FKBP4
#> SLC38A2
#> SMAGP
#> PTPN11
#> ARAP2
#> USP15
#> DCBLD1
#> MET
#> SPRED1
#> SCRIB