Calculates Spearman dissimilarity and t-SNE from a given dataset

This function computes the Spearman dissimilarity matrix from the input dataset, processes it for missing values, and performs t-SNE for dimensionality reduction.

SpearmanDissimilarity(ptmtable)

Arguments

ptmtable

A data frame containing numeric data for post-translational modifications.

Value

A matrix containing t-SNE coordinates (3D).

Examples

SpearmanDissimilarity(ptmtable)
#> Warning: NAs introduced by coercion
#> [1] "Converting Data Types..."
#> [1] "Calculating Spearman Correlation..."
#> [1] "Calculating Spearman Dissimilarity..."
#> [1] "Filtering missing values..."
#> [1] "Mapping Data Points..."
#>               [,1]       [,2]        [,3]
#>   [1,] -19.0671875  15.376244  26.4664606
#>   [2,]  12.0469946 -28.087685   6.8812322
#>   [3,]   0.8513273  17.571634 -24.6729505
#>   [4,]   9.2173707 -32.259790  16.0603515
#>   [5,]  -8.1641801  -7.789901  30.2107467
#>   [6,]  16.8344952  10.478984 -20.4775499
#>   [7,]   1.9003908  11.778344 -24.3893834
#>   [8,] -13.8815432  37.770599  11.9276775
#>   [9,]  20.5618207 -82.449652  13.9762761
#>  [10,] -26.8436610   9.318330 -31.5586549
#>  [11,]  -6.9129063  28.130790 -20.8737178
#>  [12,]  -9.1067148  -8.303364  25.3340578
#>  [13,] -22.3592487  13.643663  22.7919129
#>  [14,]  -1.8464115  28.037696 -22.5319386
#>  [15,] -16.8099004  32.047870  -6.3878395
#>  [16,]  -6.3194700  22.279895 -26.1872445
#>  [17,]  -8.2236870  28.456128  13.2729289
#>  [18,] -15.2508652  30.155779  10.3610612
#>  [19,] -25.7058233  17.201260  13.9942026
#>  [20,] -12.6596961  25.984766  -7.5385891
#>  [21,] -18.8730158  27.152003 -10.6444144
#>  [22,] -15.1723840  32.170256   1.4425605
#>  [23,] -19.8104540  26.072902   1.1705296
#>  [24,] -11.9224253 -13.024356  24.6048609
#>  [25,]  23.3381880 -70.627666  19.8746549
#>  [26,]   8.2451116  23.699504 -26.4331159
#>  [27,]   8.9432013 -76.111137  24.5418272
#>  [28,]  15.3538184 -86.660639  17.5590379
#>  [29,] -27.9036596   4.851024 -27.2394740
#>  [30,]  -0.5017386  27.585119 -17.2393479
#>  [31,] -32.9120217  32.492128   1.7173211
#>  [32,] -17.3711025  33.850450  14.0118460
#>  [33,]   2.5177144  33.416086 -12.9506457
#>  [34,] -27.1167719  19.151578   8.7652518
#>  [35,]  16.0286057 -69.727761  19.1965250
#>  [36,]  12.0949494 -75.173276  16.1181280
#>  [37,]  -7.7076880  42.230650 -12.4756948
#>  [38,]  24.5276688 -78.604344  19.1953442
#>  [39,]  -1.7690446  48.806301   6.3879574
#>  [40,]  14.8061401 -82.817754  25.1432344
#>  [41,]  14.1149862 -80.695659  12.2420696
#>  [42,]  11.2553324   4.612527 -22.2622492
#>  [43,] -15.9526523  38.303435  -7.4502045
#>  [44,]   8.8799276 -28.308087  12.9590009
#>  [45,] -22.8111327  32.736923 -12.0809541
#>  [46,]  -1.8524240  21.305473  -0.2000420
#>  [47,]  20.6261057 -84.662963  21.9081598
#>  [48,] -22.3709888  10.967698  15.1195772
#>  [49,] -17.1202648  25.132422  11.5608688
#>  [50,]  -8.8068625   6.217729  19.2336703
#>  [51,]  16.5757761  -6.776038  -9.9856555
#>  [52,]   1.5896368  11.673059  11.8190111
#>  [53,] -22.2631389  31.200276   2.7426476
#>  [54,]  16.9187134 -73.566970  25.3561716
#>  [55,]  10.9191948 -81.442412  19.2558341
#>  [56,] -22.5670961   5.866157 -36.0102826
#>  [57,] -25.4347305  27.252935  14.5625336
#>  [58,]   7.3024135  15.552713 -19.7017748
#>  [59,] -10.6591157   2.190904 -29.9431077
#>  [60,]  24.5641242  11.387076 -14.0467433
#>  [61,]  23.7910385 -25.772032  -1.3929813
#>  [62,] -27.2793342  32.828210  -4.4248476
#>  [63,]  22.3737704 -77.096535  26.5778273
#>  [64,]   9.1084904 -33.802087 -10.3810322
#>  [65,]  17.1668483 -78.833615  19.5877165
#>  [66,]  17.5582261  32.434394   4.5658022
#>  [67,]  15.0811192   8.317273  -9.7001244
#>  [68,]   8.7261818   8.629735 -16.1343530
#>  [69,]  10.6708499  11.282725 -13.5535172
#>  [70,]   0.9373026  38.899382  13.5648475
#>  [71,]  18.5176589 -13.134118   2.2388869
#>  [72,]   5.1246253  33.451365  13.1523997
#>  [73,]  -1.1636882  45.407166  14.3939798
#>  [74,]  -0.7264942  11.792980  15.1681952
#>  [75,] -14.0936587  17.344043  14.7980606
#>  [76,]  14.2543655 -33.045016  13.2894696
#>  [77,]  -8.5915264  -2.239395  21.7864993
#>  [78,]  14.2759904  20.553279 -11.6552100
#>  [79,]  17.1454022  13.555263 -16.3703025
#>  [80,]  13.8494276  33.367562   7.7123932
#>  [81,]  19.5275988 -74.970627  13.5007670
#>  [82,] -21.4173862  42.021589  -1.1379780
#>  [83,]  -4.8448904  40.271847 -34.4608434
#>  [84,]  -1.5623956  41.673235 -33.1751501
#>  [85,]  17.3143577 -25.306219  -0.3017616
#>  [86,]  17.4148674 -18.884607  -2.5990300
#>  [87,]  22.6077209 -28.794240   1.2457790
#>  [88,]   0.3933354  17.493932  -6.2336596
#>  [89,]  15.9506957  19.258512 -21.6393548
#>  [90,]  -5.3961220  47.498806  18.6757971
#>  [91,]  16.6240134 -15.644406  -8.1908909
#>  [92,]  11.8804076 -25.887127   0.7032135
#>  [93,]  16.1127352  14.771924 -35.1778219
#>  [94,]  16.3078156  18.973274 -37.2279635
#>  [95,] -26.6907572   9.604726 -35.7150242
#>  [96,] -10.6007547  38.474594  -1.3714165
#>  [97,]  -1.1124114 -14.849206  22.5681828
#>  [98,] -17.4859322  36.247903   1.9784147
#>  [99,]  16.4435610 -26.631773  -7.1310207
#> [100,] -14.1570550   3.717435 -32.0179052