TSNE matrices Example

List of of lists of dataframes of ptm positions in 3D space to show their relationships with each other based on distance.

ex.tsne.matrices

Format

List of 3 Distance metrics (Euclidean, Spearman, and the average of Spearman and Euclidean (SED)) and the clusters that form when they are run through tsne:

Euclidean

A list of 6 clusters of ptms based on their Euclidean distance and positional distance determined by tsne

Spearman

A list of 6 clusters of ptms based on their Spearman dissimilarity and positional distance determined by tsne

SED

A list of 6 clusters of ptms based on the average of their Euclidean distance and Spearman Dissimilarity and positional distance determined by tsne

Source

"created by MakeClusterList"

Examples

data(ex.tsne.matrices)
#> Warning: data set ‘ex.tsne.matrices’ not found
head(ex.tsne.matrices)
#> $Euclidean
#> $`1`
#>                          PTM.Name group
#> 1                   AARS ubi K747     1
#> 2                 ABCA1 ubi K2023     1
#> 3                 ABCC1 ubi K1502     1
#> 4                  ABCC1 ubi K280     1
#> 5                  ABCC1 ubi K303     1
#> 6                  ABCC1 ubi K498     1
#> 7                  ABCC1 ubi K636     1
#> 8                  ABCC1 ubi K814     1
#> 9                  ABCC1 ubi K891     1
#> 10                 ABCC1 ubi K894     1
#> 11                 ABCC1 ubi K908     1
#> 12                 ABCC1 ubi K936     1
#> 13                 ABCC2 ubi K947     1
#> 14                 ABCC3 ubi K931     1
#> 15                ABCC4 ubi K1289     1
#> 16 ABCD1 ubi K407; ABCD2 ubi K411     1
#> 17                 ABCD3 ubi K260     1
#> 18                 ABCD3 ubi K576     1
#> 19                 ABCE1 ubi K210     1
#> 20                 ABCE1 ubi K250     1
#> 21                  ABCE1 ubi K93     1
#> 22                 ABHD11 ack K87     1
#> 23                ABHD17A ubi K92     1
#> 24                 ABHD3 ubi K288     1
#> 25                  ABLIM1 p Y383     1
#> 26                  ABLIM1 p Y410     1
#> 27                  ABLIM1 p Y688     1
#> 28                ABLIM1 ubi K275     1
#> 29                  ABLIM3 p S493     1
#> 30              ABRAXAS1 ubi K347     1
#> 31                 ACBD3 ubi K507     1
#> 32                 ACKR3 ubi K362     1
#> 
#> $`2`
#>           PTM.Name group
#> 1   AASDH ack K544     2
#> 2    AATF ack K145     2
#> 3   ABCD3 ack K260     2
#> 4      ABI2 p Y371     2
#> 5      ABL1 p Y185     2
#> 6    ABLIM1 ubi K8     2
#> 7    ACAA2 ack K25     2
#> 8  ACADVL ack K278     2
#> 9  ACADVL ack K331     2
#> 10  ACAT1 ack K181     2
#> 
#> $`3`
#>                    PTM.Name group
#> 1         AASDHPPT ack K151     3
#> 2              ABCA3 p S683     3
#> 3            ABCA3 ubi K684     3
#> 4            ABCB6 ubi K482     3
#> 5            ABCB6 ubi K836     3
#> 6            ABCC1 ubi K290     3
#> 7            ABCC1 ubi K496     3
#> 8           ABCC3 ubi K1498     3
#> 9            ABCC3 ubi K512     3
#> 10           ABCC3 ubi K938     3
#> 11          ABCC4 ubi K1278     3
#> 12           ABCC4 ubi K515     3
#> 13           ABCC4 ubi K622     3
#> 14              ABCC5 p Y10     3
#> 15          ABHD10 ack K101     3
#> 16              ABI1 p Y198     3
#> 17            ABI2 ubi K167     3
#> 18 ABL1 p Y393; ABL2 p Y439     3
#> 19              ABL2 p Y718     3
#> 20            ABLIM1 p Y357     3
#> 21            ABLIM1 p Y373     3
#> 22            ABLIM1 p Y406     3
#> 23            ABLIM1 p Y461     3
#> 24            ABLIM3 p Y487     3
#> 25            ABLIM3 p Y538     3
#> 26               ABR p Y854     3
#> 27             ABT1 ack K97     3
#> 28             ABT1 ack K98     3
#> 29             ACAP2 p Y742     3
#> 30             ACAP2 p Y750     3
#> 31           ACAT1 ack K124     3
#> 32           ACAT1 ack K174     3
#> 33           ACAT1 ack K190     3
#> 34           ACAT1 ack K202     3
#> 35            ACLY ack K259     3
#> 
#> $`4`
#>             PTM.Name group
#> 1     ABCB6 ubi K726     4
#> 2    ABCC3 ubi K1177     4
#> 3     ABCC3 ubi K193     4
#> 4     ABCC3 ubi K244     4
#> 5     ABCC4 ubi K264     4
#> 6     ABCC4 ubi K408     4
#> 7     ABCC4 ubi K540     4
#> 8     ABCC4 ubi K695     4
#> 9      ABCC4 ubi K77     4
#> 10    ABCE1 ubi K121     4
#> 11    ABHD2 ubi K152     4
#> 12    ABHD2 ubi K282     4
#> 13       ABL1 p T392     4
#> 14       ABL1 p Y393     4
#> 15 ABRAXAS1 ack K268     4
#> 
#> $`5`
#>         PTM.Name group
#> 1 ABCB6 ubi K771     5
#> 2 ABCC3 ubi K278     5
#> 3  ABHD2 ubi K57     5
#> 4    ABI1 p Y213     5
#> 5    ABI2 p Y213     5
#> 6 ACBD3 ubi K386     5
#> 7  ACLY ack K554     5
#> 
#> $`6`
#>         PTM.Name group
#> 1 ACIN1 ack K238     6
#> 
#> 
#> $Spearman
#> $`1`
#>        PTM.Name group
#> 1 AARS ubi K747     1
#> 
#> $`2`
#>         PTM.Name group
#> 1 AASDH ack K544     2
#> 
#> $`3`
#>            PTM.Name group
#> 1 AASDHPPT ack K151     3
#> 
#> $`4`
#>         PTM.Name group
#> 1  AATF ack K145     4
#> 2 ABCD3 ack K260     4
#> 
#> $`5`
#>          PTM.Name group
#> 1 ABCA1 ubi K2023     5
#> 2 ABCC1 ubi K1502     5
#> 
#> $`6`
#>        PTM.Name group
#> 1  ABCA3 p S683     6
#> 2 ABLIM3 p Y538     6
#> 
#> $`7`
#>         PTM.Name group
#> 1 ABCA3 ubi K684     7
#> 
#> $`8`
#>         PTM.Name group
#> 1 ABCB6 ubi K482     8
#> 
#> $`9`
#>         PTM.Name group
#> 1 ABCB6 ubi K726     9
#> 
#> $`10`
#>         PTM.Name group
#> 1 ABCB6 ubi K771    10
#> 2 ACBD3 ubi K386    10
#> 
#> $`11`
#>         PTM.Name group
#> 1 ABCB6 ubi K836    11
#> 
#> $`12`
#>         PTM.Name group
#> 1 ABCC1 ubi K280    12
#> 
#> $`13`
#>         PTM.Name group
#> 1 ABCC1 ubi K290    13
#> 
#> $`14`
#>         PTM.Name group
#> 1 ABCC1 ubi K303    14
#> 
#> $`15`
#>         PTM.Name group
#> 1 ABCC1 ubi K496    15
#> 
#> $`16`
#>         PTM.Name group
#> 1 ABCC1 ubi K498    16
#> 
#> $`17`
#>         PTM.Name group
#> 1 ABCC1 ubi K636    17
#> 
#> $`18`
#>         PTM.Name group
#> 1 ABCC1 ubi K814    18
#> 
#> $`19`
#>         PTM.Name group
#> 1 ABCC1 ubi K891    19
#> 
#> $`20`
#>         PTM.Name group
#> 1 ABCC1 ubi K894    20
#> 
#> $`21`
#>         PTM.Name group
#> 1 ABCC1 ubi K908    21
#> 2 ACKR3 ubi K362    21
#> 
#> $`22`
#>         PTM.Name group
#> 1 ABCC1 ubi K936    22
#> 
#> $`23`
#>         PTM.Name group
#> 1 ABCC2 ubi K947    23
#> 
#> $`24`
#>          PTM.Name group
#> 1 ABCC3 ubi K1177    24
#> 
#> $`25`
#>          PTM.Name group
#> 1 ABCC3 ubi K1498    25
#> 
#> $`26`
#>         PTM.Name group
#> 1 ABCC3 ubi K193    26
#> 
#> $`27`
#>         PTM.Name group
#> 1 ABCC3 ubi K244    27
#> 
#> $`28`
#>         PTM.Name group
#> 1 ABCC3 ubi K278    28
#> 
#> $`29`
#>         PTM.Name group
#> 1 ABCC3 ubi K512    29
#> 
#> $`30`
#>         PTM.Name group
#> 1 ABCC3 ubi K931    30
#> 
#> $`31`
#>         PTM.Name group
#> 1 ABCC3 ubi K938    31
#> 
#> $`32`
#>          PTM.Name group
#> 1 ABCC4 ubi K1278    32
#> 
#> $`33`
#>          PTM.Name group
#> 1 ABCC4 ubi K1289    33
#> 
#> $`34`
#>         PTM.Name group
#> 1 ABCC4 ubi K264    34
#> 
#> $`35`
#>         PTM.Name group
#> 1 ABCC4 ubi K408    35
#> 
#> $`36`
#>         PTM.Name group
#> 1 ABCC4 ubi K515    36
#> 
#> $`37`
#>         PTM.Name group
#> 1 ABCC4 ubi K540    37
#> 
#> $`38`
#>         PTM.Name group
#> 1 ABCC4 ubi K622    38
#> 
#> $`39`
#>         PTM.Name group
#> 1 ABCC4 ubi K695    39
#> 
#> $`40`
#>        PTM.Name group
#> 1 ABCC4 ubi K77    40
#> 
#> $`41`
#>      PTM.Name group
#> 1 ABCC5 p Y10    41
#> 
#> $`42`
#>                         PTM.Name group
#> 1 ABCD1 ubi K407; ABCD2 ubi K411    42
#> 
#> $`43`
#>         PTM.Name group
#> 1 ABCD3 ubi K260    43
#> 
#> $`44`
#>         PTM.Name group
#> 1 ABCD3 ubi K576    44
#> 
#> $`45`
#>         PTM.Name group
#> 1 ABCE1 ubi K121    45
#> 
#> $`46`
#>         PTM.Name group
#> 1 ABCE1 ubi K210    46
#> 
#> $`47`
#>         PTM.Name group
#> 1 ABCE1 ubi K250    47
#> 
#> $`48`
#>        PTM.Name group
#> 1 ABCE1 ubi K93    48
#> 
#> $`49`
#>          PTM.Name group
#> 1 ABHD10 ack K101    49
#> 
#> $`50`
#>         PTM.Name group
#> 1 ABHD11 ack K87    50
#> 2  ABLIM1 p Y688    50
#> 
#> $`51`
#>          PTM.Name group
#> 1 ABHD17A ubi K92    51
#> 
#> $`52`
#>         PTM.Name group
#> 1 ABHD2 ubi K152    52
#> 
#> $`53`
#>         PTM.Name group
#> 1 ABHD2 ubi K282    53
#> 
#> $`54`
#>        PTM.Name group
#> 1 ABHD2 ubi K57    54
#> 
#> $`55`
#>         PTM.Name group
#> 1 ABHD3 ubi K288    55
#> 
#> $`56`
#>      PTM.Name group
#> 1 ABI1 p Y198    56
#> 
#> $`57`
#>        PTM.Name group
#> 1   ABI1 p Y213    57
#> 2 ACLY ack K554    57
#> 
#> $`58`
#>      PTM.Name group
#> 1 ABI2 p Y213    58
#> 
#> $`59`
#>          PTM.Name group
#> 1     ABI2 p Y371    59
#> 2 ACADVL ack K331    59
#> 
#> $`60`
#>        PTM.Name group
#> 1 ABI2 ubi K167    60
#> 
#> $`61`
#>      PTM.Name group
#> 1 ABL1 p T392    61
#> 
#> $`62`
#>      PTM.Name group
#> 1 ABL1 p Y185    62
#> 
#> $`63`
#>      PTM.Name group
#> 1 ABL1 p Y393    63
#> 
#> $`64`
#>                   PTM.Name group
#> 1 ABL1 p Y393; ABL2 p Y439    64
#> 2               ABR p Y854    64
#> 
#> $`65`
#>      PTM.Name group
#> 1 ABL2 p Y718    65
#> 
#> $`66`
#>        PTM.Name group
#> 1 ABLIM1 p Y357    66
#> 2 ABLIM1 p Y373    66
#> 
#> $`67`
#>        PTM.Name group
#> 1 ABLIM1 p Y383    67
#> 
#> $`68`
#>        PTM.Name group
#> 1 ABLIM1 p Y406    68
#> 
#> $`69`
#>        PTM.Name group
#> 1 ABLIM1 p Y410    69
#> 
#> $`70`
#>        PTM.Name group
#> 1 ABLIM1 p Y461    70
#> 
#> $`71`
#>          PTM.Name group
#> 1 ABLIM1 ubi K275    71
#> 
#> $`72`
#>        PTM.Name group
#> 1 ABLIM1 ubi K8    72
#> 
#> $`73`
#>        PTM.Name group
#> 1 ABLIM3 p S493    73
#> 
#> $`74`
#>        PTM.Name group
#> 1 ABLIM3 p Y487    74
#> 
#> $`75`
#>            PTM.Name group
#> 1 ABRAXAS1 ack K268    75
#> 
#> $`76`
#>            PTM.Name group
#> 1 ABRAXAS1 ubi K347    76
#> 
#> $`77`
#>       PTM.Name group
#> 1 ABT1 ack K97    77
#> 2 ABT1 ack K98    77
#> 
#> $`78`
#>        PTM.Name group
#> 1 ACAA2 ack K25    78
#> 
#> $`79`
#>          PTM.Name group
#> 1 ACADVL ack K278    79
#> 
#> $`80`
#>       PTM.Name group
#> 1 ACAP2 p Y742    80
#> 
#> $`81`
#>       PTM.Name group
#> 1 ACAP2 p Y750    81
#> 
#> $`82`
#>         PTM.Name group
#> 1 ACAT1 ack K124    82
#> 
#> $`83`
#>         PTM.Name group
#> 1 ACAT1 ack K174    83
#> 
#> $`84`
#>         PTM.Name group
#> 1 ACAT1 ack K181    84
#> 
#> $`85`
#>         PTM.Name group
#> 1 ACAT1 ack K190    85
#> 2 ACAT1 ack K202    85
#> 
#> $`86`
#>         PTM.Name group
#> 1 ACBD3 ubi K507    86
#> 
#> $`87`
#>         PTM.Name group
#> 1 ACIN1 ack K238    87
#> 
#> $`88`
#>        PTM.Name group
#> 1 ACLY ack K259    88
#> 
#> 
#> $SED
#> $`1`
#>                          PTM.Name group
#> 1                   AARS ubi K747     1
#> 2                 ABCA1 ubi K2023     1
#> 3                 ABCC1 ubi K1502     1
#> 4                  ABCC1 ubi K280     1
#> 5                  ABCC1 ubi K303     1
#> 6                  ABCC1 ubi K498     1
#> 7                  ABCC1 ubi K636     1
#> 8                  ABCC1 ubi K814     1
#> 9                  ABCC1 ubi K891     1
#> 10                 ABCC1 ubi K894     1
#> 11                 ABCC1 ubi K908     1
#> 12                 ABCC1 ubi K936     1
#> 13                 ABCC2 ubi K947     1
#> 14                 ABCC3 ubi K931     1
#> 15                ABCC4 ubi K1289     1
#> 16 ABCD1 ubi K407; ABCD2 ubi K411     1
#> 17                 ABCD3 ubi K260     1
#> 18                 ABCD3 ubi K576     1
#> 19                 ABCE1 ubi K210     1
#> 20                 ABCE1 ubi K250     1
#> 21                  ABCE1 ubi K93     1
#> 22                 ABHD11 ack K87     1
#> 23                ABHD17A ubi K92     1
#> 24                 ABHD3 ubi K288     1
#> 25                  ABLIM1 p Y383     1
#> 26                  ABLIM1 p Y410     1
#> 27                  ABLIM1 p Y688     1
#> 28                ABLIM1 ubi K275     1
#> 29                  ABLIM3 p S493     1
#> 30              ABRAXAS1 ubi K347     1
#> 31                 ACBD3 ubi K507     1
#> 32                 ACKR3 ubi K362     1
#> 
#> $`2`
#>           PTM.Name group
#> 1   AASDH ack K544     2
#> 2    AATF ack K145     2
#> 3   ABCD3 ack K260     2
#> 4      ABI2 p Y371     2
#> 5      ABL1 p Y185     2
#> 6    ABLIM1 ubi K8     2
#> 7    ACAA2 ack K25     2
#> 8  ACADVL ack K278     2
#> 9  ACADVL ack K331     2
#> 10  ACAT1 ack K181     2
#> 
#> $`3`
#>                    PTM.Name group
#> 1         AASDHPPT ack K151     3
#> 2              ABCA3 p S683     3
#> 3            ABCA3 ubi K684     3
#> 4            ABCB6 ubi K482     3
#> 5            ABCB6 ubi K836     3
#> 6            ABCC1 ubi K290     3
#> 7            ABCC1 ubi K496     3
#> 8           ABCC3 ubi K1498     3
#> 9            ABCC3 ubi K512     3
#> 10           ABCC3 ubi K938     3
#> 11          ABCC4 ubi K1278     3
#> 12           ABCC4 ubi K515     3
#> 13           ABCC4 ubi K622     3
#> 14              ABCC5 p Y10     3
#> 15          ABHD10 ack K101     3
#> 16              ABI1 p Y198     3
#> 17            ABI2 ubi K167     3
#> 18 ABL1 p Y393; ABL2 p Y439     3
#> 19              ABL2 p Y718     3
#> 20            ABLIM1 p Y357     3
#> 21            ABLIM1 p Y373     3
#> 22            ABLIM1 p Y406     3
#> 23            ABLIM1 p Y461     3
#> 24            ABLIM3 p Y487     3
#> 25            ABLIM3 p Y538     3
#> 26               ABR p Y854     3
#> 27             ABT1 ack K97     3
#> 28             ABT1 ack K98     3
#> 29             ACAP2 p Y742     3
#> 30             ACAP2 p Y750     3
#> 31           ACAT1 ack K124     3
#> 32           ACAT1 ack K174     3
#> 33           ACAT1 ack K190     3
#> 34           ACAT1 ack K202     3
#> 35            ACLY ack K259     3
#> 
#> $`4`
#>             PTM.Name group
#> 1     ABCB6 ubi K726     4
#> 2    ABCC3 ubi K1177     4
#> 3     ABCC3 ubi K193     4
#> 4     ABCC3 ubi K244     4
#> 5     ABCC4 ubi K264     4
#> 6     ABCC4 ubi K408     4
#> 7     ABCC4 ubi K540     4
#> 8     ABCC4 ubi K695     4
#> 9      ABCC4 ubi K77     4
#> 10    ABCE1 ubi K121     4
#> 11    ABHD2 ubi K152     4
#> 12    ABHD2 ubi K282     4
#> 13       ABL1 p T392     4
#> 14       ABL1 p Y393     4
#> 15 ABRAXAS1 ack K268     4
#> 
#> $`5`
#>         PTM.Name group
#> 1 ABCB6 ubi K771     5
#> 2 ABCC3 ubi K278     5
#> 3  ABHD2 ubi K57     5
#> 4    ABI1 p Y213     5
#> 5    ABI2 p Y213     5
#> 6 ACBD3 ubi K386     5
#> 7  ACLY ack K554     5
#> 
#> $`6`
#>         PTM.Name group
#> 1 ACIN1 ack K238     6
#> 
#>