Calculates Spearman dissimilarity and t-SNE from a given dataset

This function computes the Spearman dissimilarity matrix from the input dataset, processes it for missing values, and performs t-SNE for dimensionality reduction.

SpearmanDissimilarity(ptmtable)

Arguments

ptmtable

A data frame containing numeric data for post-translational modifications.

Value

A matrix containing t-SNE coordinates (3D).

Examples

SpearmanDissimilarity(ptmtable)
#> Warning: NAs introduced by coercion
#> [1] "Converting Data Types..."
#> [1] "Calculating Spearman Correlation..."
#> [1] "Calculating Spearman Dissimilarity..."
#> [1] "Filtering missing values..."
#> [1] "Mapping Data Points..."
#>               [,1]          [,2]         [,3]
#>   [1,]  23.7102285 -26.530966138   1.71759227
#>   [2,]  -7.9834830   9.934470361  26.92423925
#>   [3,] -10.0740925   3.523988931 -29.91047432
#>   [4,]  -3.5251417   5.717193534  34.10954546
#>   [5,]  17.4179650 -11.343949920  19.51799035
#>   [6,] -13.7253573  13.693686214 -14.55424875
#>   [7,] -10.5225171   8.498194096 -28.38640955
#>   [8,]   9.4451656 -24.438874107 -21.63864520
#>   [9,]  -9.5324355  33.161891779  67.62492102
#>  [10,]  13.9434119  19.978515343 -40.37750861
#>  [11,]  -2.8109306  -6.659346107 -32.95153248
#>  [12,]  13.1649445 -12.919190057  19.09409745
#>  [13,]  24.5339358 -22.267131651  -0.04136903
#>  [14,]  -7.0761835  -5.775266584 -30.86422184
#>  [15,]  11.9949316 -10.970461286 -24.55074825
#>  [16,]  -5.7231006  -0.048082704 -35.26959604
#>  [17,]  23.2690422 -24.489936261 -20.86401046
#>  [18,]  16.2025684 -22.750336889 -19.09734393
#>  [19,]  24.7200302 -19.064839880  -7.37656885
#>  [20,]   5.4949876 -14.046141541 -23.80671188
#>  [21,]   6.9560119 -13.634536363 -29.58369221
#>  [22,]  21.3275831 -13.146578656 -23.07315001
#>  [23,]  16.3011500 -15.472515420 -20.70047711
#>  [24,]  11.5893970 -14.869710411  23.50982321
#>  [25,] -16.4011885  20.806749847  65.77499448
#>  [26,] -14.3900958  -2.232095706 -24.77242031
#>  [27,]  -2.6007989  23.195713802  70.63865183
#>  [28,] -14.7215170  27.752903449  76.51440928
#>  [29,]  16.4852962  22.997451753 -36.15279266
#>  [30,]  -5.3146220  -8.002861932 -26.88344632
#>  [31,]  16.8042470 -24.986288998 -34.67135315
#>  [32,]  13.4381704 -27.033262882 -19.47740365
#>  [33,]  -6.6299059 -14.679316091 -25.15877085
#>  [34,]  26.0472529 -16.898810893 -12.04798270
#>  [35,]  -9.8686744  21.446300184  64.46855604
#>  [36,]  -7.1842056  27.551369213  65.40881230
#>  [37,]  13.1018009  -1.253948317 -34.15651818
#>  [38,] -18.1632614  26.151380715  70.84625722
#>  [39,]  -2.6298331 -31.725102296 -28.93489645
#>  [40,]  -8.2669698  25.776339888  76.29050877
#>  [41,] -11.4778035  27.125019117  70.25756044
#>  [42,] -18.0024795  10.187166331 -18.88026748
#>  [43,]  14.7362613  -5.254691503 -25.44763747
#>  [44,]  -3.8376939   5.975642702  29.52968832
#>  [45,]  10.5192613 -12.202328558 -34.27165892
#>  [46,]  10.2271383  -2.290317962 -16.23685974
#>  [47,]  -9.7303382  32.298823233  75.20367862
#>  [48,]  21.2296363 -16.111671352  -2.53306397
#>  [49,]  15.9439745 -21.062349148 -14.06187372
#>  [50,]  10.8010595 -16.268505252   5.48722093
#>  [51,] -14.5099811   9.618701628   2.57217868
#>  [52,]   3.0240198 -10.736912320  -3.22920837
#>  [53,]  16.7482307 -18.737318972 -25.57731203
#>  [54,]  -9.6603977  19.993123083  71.01025204
#>  [55,]  -5.5666095  29.384911419  71.49771331
#>  [56,]   8.3141396  23.082695734 -40.76953425
#>  [57,]  20.1851917 -26.689278281 -12.20192612
#>  [58,] -10.6736396   4.628501482 -21.49456743
#>  [59,]   2.1930774  21.486074142 -29.88732563
#>  [60,]  -5.9226279  18.669885951 -13.13275294
#>  [61,] -18.8990192  15.562080970  21.17040608
#>  [62,]  13.6186461 -18.953412652 -32.80041964
#>  [63,] -15.0564402  20.207753187  73.77844097
#>  [64,] -11.0756371  25.649507528  22.98286188
#>  [65,] -15.0968424  32.414382019  71.16536782
#>  [66,] -15.2423874 -22.100481799 -12.86415298
#>  [67,]  -9.4220860   7.454348540  -8.59393807
#>  [68,] -12.7020538   4.969372157 -15.06504447
#>  [69,]  -9.0647747   6.237973653 -14.44681839
#>  [70,]  -0.3167209 -30.270215881 -17.50311640
#>  [71,]  -6.8692514  16.150329689  11.51916993
#>  [72,]  -3.0518493 -27.434939954 -12.60352770
#>  [73,]   0.5080587 -34.187984031 -22.31853519
#>  [74,]   4.6046780 -12.699984246  -0.48941035
#>  [75,]  14.1057566 -19.559045591  -6.16920792
#>  [76,]  -8.6782350   6.991973927  33.85466851
#>  [77,]  11.2292389 -14.327968831  13.11331675
#>  [78,] -11.9196608  16.292154411 -23.31757761
#>  [79,]  -9.6585074  13.039022819 -16.36492993
#>  [80,] -11.3898768 -24.455143802 -12.97028450
#>  [81,] -14.2635619  28.326565020  64.63642733
#>  [82,]  26.6015605 -11.049619757 -28.11490525
#>  [83,] -14.1749975  -8.560642250 -45.62283939
#>  [84,] -11.3776827  -7.529025402 -47.36684586
#>  [85,] -13.8006841  12.474687139  20.57838766
#>  [86,] -12.8028590  14.600700431  14.67113391
#>  [87,] -18.1657496  13.792521064  24.47575642
#>  [88,]   6.9949878   2.495801299 -19.00944288
#>  [89,]  -4.5152499   9.776866544 -19.52180397
#>  [90,]   5.3183613 -37.447171854 -22.86834402
#>  [91,] -16.0475781  14.590328136   9.52625749
#>  [92,]  -9.1164728  11.141205342  21.42001988
#>  [93,] -24.4095871  -0.006865891 -16.69365929
#>  [94,] -25.8586281  -3.138918049 -19.19896483
#>  [95,]  11.7573614  20.453603253 -43.44818822
#>  [96,]  17.8905795  -9.150164341 -30.00410776
#>  [97,]   8.3514336  -5.131091078  22.46371827
#>  [98,]  21.2863218 -14.461074171 -26.99925879
#>  [99,] -16.6129326   7.946244427  17.42151905
#> [100,]   3.9781157  21.852558535 -33.39453897