Calculates Spearman dissimilarity and t-SNE from a given dataset

This function computes the Spearman dissimilarity matrix from the input dataset, processes it for missing values, and performs t-SNE for dimensionality reduction.

SpearmanDissimilarity(ptmtable)

Arguments

ptmtable

A data frame containing numeric data for post-translational modifications.

Value

A matrix containing t-SNE coordinates (3D).

Examples

SpearmanDissimilarity(ptmtable)
#> [1] "Converting Data Types..."
#> [1] "Calculating Spearman Correlation..."
#> [1] "Calculating Spearman Dissimilarity..."
#> [1] "Filtering missing values..."
#> [1] "Mapping Data Points..."
#>               [,1]        [,2]        [,3]
#>   [1,]  -7.3821133  24.9630096  10.3361643
#>   [2,]   0.4056240 -17.5327881  19.0206884
#>   [3,]   5.2461372   0.9448896 -26.9042886
#>   [4,]  -2.2793894 -14.4819602  26.5290942
#>   [5,]  -4.4561556  11.4803901  20.1316055
#>   [6,]   5.0780800 -12.0751539 -19.3704919
#>   [7,]   1.1587250  -1.1462237 -26.2874425
#>   [8,] -13.0147087  19.4267785 -14.0468365
#>   [9,]  10.6034208 -34.3678486  56.8687667
#>  [10,]  28.5444657  16.0277312 -17.9931591
#>  [11,]   7.7332782  11.7865742 -25.3230099
#>  [12,]  -7.6294730   9.3010393  18.8451105
#>  [13,]  -6.5347340  25.5337189   6.3682015
#>  [14,]   9.2842673   7.9708524 -24.3234948
#>  [15,]  -1.4002736  21.6640459 -20.8668654
#>  [16,]   6.5499869   6.9858575 -29.9060163
#>  [17,] -13.2510761  31.0483785 -14.2117844
#>  [18,] -11.1276235  25.2269034 -12.6524478
#>  [19,]  -5.5521611  27.4435048  -0.1208013
#>  [20,]   0.7079246  17.5909657 -16.9487172
#>  [21,]   4.2770687  21.3400993 -19.2965010
#>  [22,]  -6.5345246  24.6837604 -18.3901700
#>  [23,]  -2.5855754  24.7554442 -13.2016812
#>  [24,] -10.2037317   7.9241226  22.3118609
#>  [25,]  -3.2450232 -35.5840168  55.7725962
#>  [26,]  11.3652081  -2.9061708 -25.6465954
#>  [27,]   7.0964698 -29.1163504  57.7073191
#>  [28,]   5.4739381 -43.4540158  59.2991881
#>  [29,]  28.3765345  15.7402154 -12.8561198
#>  [30,]   6.6105395   8.0150184 -20.6131481
#>  [31,]   2.0154723  35.0912764 -15.4630214
#>  [32,] -15.2349931  23.3852061 -13.2927006
#>  [33,]   7.8674251  10.4739154 -14.3688401
#>  [34,]  -2.2716295  27.9351572  -3.5048968
#>  [35,]   0.6145110 -32.3616060  59.1123881
#>  [36,]  10.7164927 -39.8122732  54.8527234
#>  [37,]  -4.9326445  16.1211652 -31.2783183
#>  [38,]   0.4494449 -40.8295404  56.8937589
#>  [39,] -21.0463148  17.8680936 -26.4943812
#>  [40,]   5.2776891 -36.7909525  57.2965874
#>  [41,]   3.1470214 -38.8100148  62.4094161
#>  [42,]   4.7176299 -10.4986051 -12.6940139
#>  [43,]  -3.4651878  22.2267794 -25.8178375
#>  [44,]  -1.4519186 -13.6639589  22.6024380
#>  [45,]   4.2854707  24.7285244 -23.8482678
#>  [46,]  -8.2855316  10.4052563 -17.1444764
#>  [47,]   7.1944309 -34.5451201  62.1158649
#>  [48,] -11.7944019  26.0168054   2.0735754
#>  [49,]  -9.9861642  24.0788056  -8.3542165
#>  [50,] -10.4734843  12.8573921   6.9414058
#>  [51,]   1.0353103 -17.3857445  -3.3472155
#>  [52,]  -9.5875096   7.0580794  -2.3856832
#>  [53,]  -3.7377594  28.6719001 -15.6323991
#>  [54,]   2.4133331 -32.5238068  53.3552914
#>  [55,]   9.4341719 -39.4041490  60.3594275
#>  [56,]  30.8103676  10.9048178 -19.3591112
#>  [57,]  -9.3516164  30.8510015  -6.2534486
#>  [58,]   5.2960707  -2.8655343 -20.3799935
#>  [59,]  23.8642470   2.6371560 -16.0372318
#>  [60,]  -3.4079061 -15.7967011 -20.5829492
#>  [61,]  -0.3177518 -26.8155948  11.6881609
#>  [62,]   2.2373689  29.8350167 -19.6252570
#>  [63,]   0.8334439 -37.6670127  50.9935026
#>  [64,]  13.7214833 -23.1651386  15.8996303
#>  [65,]   5.6542056 -41.6972276  53.7515599
#>  [66,] -19.8101023  -2.3781495 -19.6855997
#>  [67,]  -4.1067498  -8.9419263 -15.0807030
#>  [68,]   1.5178404  -5.2093499 -15.9220570
#>  [69,]  -0.7762763  -6.0546568 -18.2796167
#>  [70,] -22.3257921  13.5768740 -17.2987954
#>  [71,]  -5.0129824 -17.9776046   6.1108175
#>  [72,] -22.0743385   8.6328680 -14.9020723
#>  [73,] -24.6117658  17.5309467 -20.3087662
#>  [74,] -11.8802550   8.1937421  -0.2910479
#>  [75,] -10.8128240  20.1243938  -2.2178618
#>  [76,]  -2.9950918 -18.8446853  24.8537969
#>  [77,]  -9.2644170  10.1194354  13.3617626
#>  [78,]  10.3993501  -6.5208112 -16.3693416
#>  [79,]   1.2213090 -11.2462755 -19.9208414
#>  [80,] -21.2351177   1.1321712 -18.2905399
#>  [81,]   7.5324131 -35.7613429  51.6661503
#>  [82,]  -5.7051674  30.2500686 -25.2229222
#>  [83,]  11.1959437  12.4439382 -40.9130191
#>  [84,]   8.7669879  10.9914202 -42.0747466
#>  [85,]   1.5906181 -21.9570657  12.9005412
#>  [86,]   0.5431200 -20.3081553   7.7054546
#>  [87,]  -0.3998804 -26.3910927  15.0439854
#>  [88,]  -5.3511831   5.8245363 -19.8920656
#>  [89,]   1.4696141  -9.2719208 -25.7888079
#>  [90,] -26.1889267  22.1168281 -18.4060268
#>  [91,]   3.2842440 -19.7052548   3.1817095
#>  [92,]   3.4519371 -18.2358825  14.6856743
#>  [93,]  13.3627335 -13.6618905 -27.3138533
#>  [94,]  16.3042186 -11.6595663 -28.5858549
#>  [95,]  30.6003482  15.2757148 -20.5583795
#>  [96,]  -9.2671430  21.5154425 -23.8560269
#>  [97,]  -3.8810414   1.5045965  20.1696501
#>  [98,]  -7.5935980  27.6127349 -20.3551032
#>  [99,]   6.5552505 -23.5412945  10.5639051
#> [100,]  25.9408425   5.1190743 -17.4218941