Calculates Spearman dissimilarity and t-SNE from a given dataset

This function computes the Spearman dissimilarity matrix from the input dataset, processes it for missing values, and performs t-SNE for dimensionality reduction.

SpearmanDissimilarity(ptmtable)

Arguments

ptmtable

A data frame containing numeric data for post-translational modifications.

Value

A matrix containing t-SNE coordinates (3D).

Examples

SpearmanDissimilarity(ptmtable)
#> Warning: NAs introduced by coercion
#> [1] "Converting Data Types..."
#> [1] "Calculating Spearman Correlation..."
#> [1] "Calculating Spearman Dissimilarity..."
#> [1] "Filtering missing values..."
#> [1] "Mapping Data Points..."
#>               [,1]        [,2]        [,3]
#>   [1,] -27.9768664 -17.1792111 -25.4504362
#>   [2,]  13.9100247  26.5827919  -7.7011808
#>   [3,]   5.2127760 -23.2793236  24.6894028
#>   [4,]   9.9028964  31.6371555 -15.6701720
#>   [5,] -17.2996119  13.4733136 -20.8285157
#>   [6,]  18.9202771 -13.8311712  16.9085099
#>   [7,]   8.6579157 -19.0711006  26.6654166
#>   [8,] -13.8148635 -28.8000262 -10.2553915
#>   [9,]  22.6426939  77.7622997 -13.9740089
#>  [10,] -21.1146559 -17.6404199  41.4001463
#>  [11,]  -8.4399379 -27.6241551  20.8023098
#>  [12,] -13.6454233  11.1565710 -23.0903978
#>  [13,] -30.2264817 -14.1510643 -22.0086891
#>  [14,]  -4.6393247 -24.3836013  20.3641589
#>  [15,] -23.7093730 -29.7642532   6.7547718
#>  [16,]  -2.2243102 -27.3805028  27.1860642
#>  [17,] -24.4473732 -30.0817949 -13.1386732
#>  [18,] -20.3195818 -25.1476967  -8.6937734
#>  [19,] -31.6536056 -14.8637600 -13.1842655
#>  [20,] -21.5910924 -33.8083971  12.1169610
#>  [21,] -21.6017153 -27.3031153  13.2919441
#>  [22,] -26.3031021 -32.2973353  -1.5961498
#>  [23,] -29.8213317 -26.5366972  -0.6125149
#>  [24,] -11.8474076  13.8776547 -27.4510296
#>  [25,]  22.7304524  65.3790635 -20.9312007
#>  [26,]  -0.1611952 -15.5993010  20.5793438
#>  [27,]  17.2554120  72.7901407 -24.2789770
#>  [28,]  19.6794672  79.6273281 -20.5208051
#>  [29,] -23.3355405 -12.3803240  39.0689504
#>  [30,]  -2.3270996 -27.3175096  16.4848992
#>  [31,] -25.5058214 -13.8853326   3.7835003
#>  [32,] -14.8144021 -23.6748604  -8.2564869
#>  [33,]  -0.4469988 -33.0019393  12.3396370
#>  [34,] -33.8282714 -18.7744254  -9.5301711
#>  [35,]  29.2169988  66.8916276 -19.6735563
#>  [36,]  20.4928140  72.0709606 -16.9949060
#>  [37,] -10.8121997 -38.3745942   8.8461258
#>  [38,]  32.8042198  74.0279823 -19.0295368
#>  [39,]   2.2155502 -30.1040096 -13.8153418
#>  [40,]  30.6137356  79.8504339 -22.8721879
#>  [41,]  29.3078580  78.5083977 -14.5488034
#>  [42,]  15.4077628 -10.1563341  22.6496151
#>  [43,] -15.4235941 -27.4541824   4.7250764
#>  [44,]   9.5241250  27.7663288 -12.6285295
#>  [45,] -17.8371144 -23.0376395  11.9169115
#>  [46,]  -7.3633809 -19.3902175  -0.1240671
#>  [47,]  25.7910940  82.5546119 -19.3469268
#>  [48,] -28.6645842  -8.8167273 -16.3986690
#>  [49,] -21.4192846 -19.8224043  -9.9832236
#>  [50,] -14.1086062  -3.8196756 -19.6177954
#>  [51,]  18.2265435   4.4546528   7.7478157
#>  [52,]  -3.8864469  -8.3569523 -11.6346635
#>  [53,] -24.1781142 -25.2322312  -1.2368948
#>  [54,]  30.6935080  71.3914909 -25.5502636
#>  [55,]  25.3895943  74.7713815 -20.5776291
#>  [56,] -15.8004362 -15.3942328  45.2975546
#>  [57,] -24.5280690 -13.8534492  -5.4685444
#>  [58,]   7.7461493 -17.0022802  18.6915029
#>  [59,]  -5.4678277 -10.9908142  38.0553128
#>  [60,]  24.3020177 -18.0073807  10.2996709
#>  [61,]  26.3547404  21.6343690  -2.5322332
#>  [62,] -21.3100028 -19.9936644   5.9195780
#>  [63,]  24.4099178  70.2737264 -27.4045010
#>  [64,]  18.5631893  33.4797930   8.2670721
#>  [65,]  24.4943904  78.5606636 -25.8946512
#>  [66,]  10.4404454 -42.7471726  -0.1576226
#>  [67,]  15.0703082 -12.6226732   7.6336095
#>  [68,]  10.6433525 -10.8517127  15.6720916
#>  [69,]  11.0654049 -13.3752398  12.4797746
#>  [70,]  -5.1080104 -39.0637893 -12.1832148
#>  [71,]  18.4078734  10.6581567  -4.7081421
#>  [72,]  -1.1023800 -42.4307400  -8.3461670
#>  [73,]  -2.9173349 -35.1018253 -17.2395546
#>  [74,]  -5.8669439  -8.2598121 -15.1241788
#>  [75,] -19.6157498 -13.0181870 -14.7439072
#>  [76,]  15.4615042  30.9093828 -14.5913428
#>  [77,] -13.2005356   4.3436313 -21.9155371
#>  [78,]  13.2350429 -24.3726130  12.7850816
#>  [79,]  17.2952927 -17.7379068  14.3461469
#>  [80,]   6.7026303 -43.0956605  -3.2344745
#>  [81,]  27.1622596  71.4445849 -13.9574016
#>  [82,] -25.8238461 -38.8285061   1.3091185
#>  [83,]  -1.4046405 -44.9300948  28.1129540
#>  [84,]  -4.9780128 -44.4778586  29.0419195
#>  [85,]  19.9786420  22.5515324  -2.0950785
#>  [86,]  19.2212584  16.2981379   0.1243969
#>  [87,]  25.2551704  24.7553313  -4.9594608
#>  [88,]  -2.3109727 -17.8725793   4.9213798
#>  [89,]  17.4827987 -20.8355390  21.7835397
#>  [90,]  -5.9330257 -33.2290220 -22.2277311
#>  [91,]  18.2087501  13.2058037   5.6599287
#>  [92,]  14.8566195  24.3898715  -1.7530392
#>  [93,]   5.4468701  -0.8801885  20.0627561
#>  [94,]   1.1487121   0.5779490  20.7266951
#>  [95,] -19.8082578 -19.0826113  44.9943824
#>  [96,] -15.4614030 -33.4220074   0.4413044
#>  [97,]  -6.8148602  17.5097023 -17.9954436
#>  [98,] -21.9593198 -32.1457059  -2.2908298
#>  [99,]  21.0650124  23.7004293   4.6251759
#> [100,]  -8.4137305 -12.9016986  40.4563828